+++ data-base-app/src/test/java/by/ita/chernook/OrderServiceTest.java
@@ -14,6 +14,7 @@ import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 
+import java.math.BigDecimal;
 import java.util.*;
 
 import static org.junit.jupiter.api.Assertions.*;
@@ -31,7 +32,7 @@ class OrderServiceTest extends TestUtils {
     @Test
     void insertOrder_then_return() {
         Order testOrder = buildOrder(UUID.randomUUID(), List.of(new OrderItem()), new DeliveryAddress(), new Customer(),
-                OrderStatusEnum.NEW, "Test comment",25.2);
+                OrderStatusEnum.NEW, "Test comment", BigDecimal.valueOf(25.2));
 
         when(orderRepository.save(testOrder)).thenReturn(testOrder);
 
@@ -46,7 +47,7 @@ class OrderServiceTest extends TestUtils {
     void updateOrder_then_return() {
         UUID orderUuid = UUID.randomUUID();
         Order testOrder = buildOrder(orderUuid, List.of(new OrderItem()), new DeliveryAddress(), new Customer(),
-                OrderStatusEnum.NEW, "Test comment",25.2);
+                OrderStatusEnum.NEW, "Test comment", BigDecimal.valueOf(25.2));
 
         when(orderRepository.existsById(orderUuid)).thenReturn(true);
         when(orderRepository.save(testOrder)).thenReturn(testOrder);
@@ -63,7 +64,7 @@ class OrderServiceTest extends TestUtils {
     void updateOrder_then_throws_NoSuchElementException() {
         UUID orderUuid = UUID.randomUUID();
         Order testOrder = buildOrder(orderUuid, List.of(new OrderItem()), new DeliveryAddress(), new Customer(),
-                OrderStatusEnum.NEW, "Test comment",55.2);
+                OrderStatusEnum.NEW, "Test comment", BigDecimal.valueOf(55.2));
 
         when(orderRepository.existsById(orderUuid)).thenReturn(false);
 
@@ -77,7 +78,7 @@ class OrderServiceTest extends TestUtils {
     void deleteOrder_then_return() {
         UUID orderUuid = UUID.randomUUID();
         Order testOrder = buildOrder(orderUuid, List.of(new OrderItem()), new DeliveryAddress(), new Customer(),
-                OrderStatusEnum.NEW, "Test comment",55.2);
+                OrderStatusEnum.NEW, "Test comment", BigDecimal.valueOf(55.2));
 
         when(orderRepository.findById(orderUuid)).thenReturn(Optional.of(testOrder));
         doNothing().when(orderRepository).deleteById(orderUuid);
@@ -106,7 +107,7 @@ class OrderServiceTest extends TestUtils {
     void findOrderById_then_return() {
         UUID orderUuid = UUID.randomUUID();
         Order testOrder = buildOrder(orderUuid, List.of(new OrderItem()), new DeliveryAddress(), new Customer(),
-                OrderStatusEnum.NEW, "Test comment",25.2);
+                OrderStatusEnum.NEW, "Test comment", BigDecimal.valueOf(25.2));
 
         when(orderRepository.findById(orderUuid)).thenReturn(Optional.of(testOrder));
 
@@ -131,9 +132,9 @@ class OrderServiceTest extends TestUtils {
     @Test
     void findAll_then_return() {
         Order testOrder = buildOrder(UUID.randomUUID(), List.of(new OrderItem()), new DeliveryAddress(), new Customer(),
-                OrderStatusEnum.PROCESSING, "Test comment",25.2);
+                OrderStatusEnum.PROCESSING, "Test comment", BigDecimal.valueOf(25.2));
         Order testOrder2 = buildOrder(UUID.randomUUID(), List.of(new OrderItem()), new DeliveryAddress(), new Customer(),
-                OrderStatusEnum.NEW, "Test comment2",48.8);
+                OrderStatusEnum.NEW, "Test comment2", BigDecimal.valueOf(48.8));
         List<Order> testList = new ArrayList<>(Arrays.asList(testOrder, testOrder2));
 
         when(orderRepository.findAll()).thenReturn(testList);
@@ -149,9 +150,9 @@ class OrderServiceTest extends TestUtils {
     void findAlLByCustomerUuid_then_return() {
         UUID customerUUid = UUID.randomUUID();
         Order testOrder = buildOrder(UUID.randomUUID(), List.of(new OrderItem()), new DeliveryAddress(), new Customer(),
-                OrderStatusEnum.PROCESSING, "Test comment",25.2);
+                OrderStatusEnum.PROCESSING, "Test comment", BigDecimal.valueOf(25.2));
         Order testOrder2 = buildOrder(UUID.randomUUID(), List.of(new OrderItem()), new DeliveryAddress(), new Customer(),
-                OrderStatusEnum.NEW, "Test comment2",48.8);
+                OrderStatusEnum.NEW, "Test comment2", BigDecimal.valueOf(48.8));
         List<Order> testList = new ArrayList<>(Arrays.asList(testOrder, testOrder2));
 
         when(orderRepository.findAllByCustomerUuid(customerUUid)).thenReturn(testList);
